// apps/backend/generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const AccountType = {
  CHECKING: 'CHECKING',
  SAVINGS: 'SAVINGS',
  CREDIT_CARD: 'CREDIT_CARD',
  CASH: 'CASH'
} as const

export type AccountType = (typeof AccountType)[keyof typeof AccountType]


export const Currency = {
  BRL: 'BRL'
} as const

export type Currency = (typeof Currency)[keyof typeof Currency]


export const InvoiceStatus = {
  OPEN: 'OPEN',
  CLOSED: 'CLOSED'
} as const

export type InvoiceStatus = (typeof InvoiceStatus)[keyof typeof InvoiceStatus]


export const InvoicePaymentStatus = {
  NOT_PAID: 'NOT_PAID',
  PARTIALLY_PAID: 'PARTIALLY_PAID',
  PAID: 'PAID'
} as const

export type InvoicePaymentStatus = (typeof InvoicePaymentStatus)[keyof typeof InvoicePaymentStatus]


export const TransactionType = {
  EXPENSE: 'EXPENSE',
  INCOME: 'INCOME',
  TRANSFER: 'TRANSFER'
} as const

export type TransactionType = (typeof TransactionType)[keyof typeof TransactionType]


export const TransactionStatus = {
  PENDING: 'PENDING',
  COMPLETED: 'COMPLETED',
  FAILED: 'FAILED'
} as const

export type TransactionStatus = (typeof TransactionStatus)[keyof typeof TransactionStatus]


export const NotificationType = {
  BUDGET_WARNING: 'BUDGET_WARNING',
  BUDGET_EXCEEDED: 'BUDGET_EXCEEDED',
  RECURRENCE_DUE_WARNING: 'RECURRENCE_DUE_WARNING',
  RECURRENCE_DUE: 'RECURRENCE_DUE',
  INVOICE_CLOSING_WARNING: 'INVOICE_CLOSING_WARNING',
  INVOICE_CLOSING: 'INVOICE_CLOSING',
  INVOICE_DUE_WARNING: 'INVOICE_DUE_WARNING',
  INVOICE_DUE: 'INVOICE_DUE',
  INVOICE_OVERDUE: 'INVOICE_OVERDUE',
  INVOICE_PAID: 'INVOICE_PAID',
  INVOICE_PARTIALLY_PAID: 'INVOICE_PARTIALLY_PAID'
} as const

export type NotificationType = (typeof NotificationType)[keyof typeof NotificationType]


export const TableName = {
  CATEGORIES: 'CATEGORIES',
  SUB_CATEGORIES: 'SUB_CATEGORIES',
  INSTITUTIONS: 'INSTITUTIONS',
  BUDGETS: 'BUDGETS',
  ACCOUNTS: 'ACCOUNTS',
  INVOICES: 'INVOICES',
  RECURRENCES: 'RECURRENCES',
  TRANSACTIONS: 'TRANSACTIONS',
  NOTIFICATIONS: 'NOTIFICATIONS'
} as const

export type TableName = (typeof TableName)[keyof typeof TableName]